backend/requirements.txt
flask
flask-cors
spotipy
python-dotenv
Flask
Flask-Cors
spotipy
python-dotenv
gunicorn
